requirements.txt
todoist-python
xmlschema
bs4
pyqt5
lxml-to-dict
lxml